requirements.txt
opencv-python
numpy
matplotlib
pyzbar
Pillow
